requirements.txt
pandas==1.4.4
pyodbc==4.0.35
snowflake_connector_python==2.7.11
